requirements.txt
# Основные зависимости для Meminto
# Python >= 3.11 требуется

# Аудио обработка
torch>=2.0.0
torchaudio>=2.2.2

# CLI
click>=8.1.7

# Диаризация и транскрипция
pyannote-audio>=3.1.1
pyannote-pipeline>=3.0.1
transformers>=4.39.3
accelerate>=1.3.0

# LLM токенизация
tiktoken>=0.6.0

# Утилиты
python-dotenv>=1.0.1
requests>=2.31.0

# Для разработки (опционально)
# ruff>=0.4.2
# mypy>=1.9.0
# types-requests>=2.31.0.20240406
